+++ src/termline.c
@@ -6,6 +6,7 @@
 #include "termpriv.h"
 #include "winpriv.h"  // disable_bidi
 
+
 termline *
 newline(int cols, int bce)
 {
@@ -806,12 +807,12 @@ term_bidi_cache_hit(int line, termchar *lbefore, int width)
     if (!termchars_equal(term.pre_bidi_cache[line].chars + i, lbefore + i))
       return false;     /* line doesn't match cache */
 
-  return true;  /* it didn't match. */
+  return true;  /* all termchars matched */
 }
 
 static void
 term_bidi_cache_store(int line, termchar *lbefore, termchar *lafter,
-                      bidi_char *wcTo, int width, int size)
+                      bidi_char *wcTo, int width, int size, int bidisize)
 {
   int i;
 
@@ -846,14 +847,22 @@ term_bidi_cache_store(int line, termchar *lbefore, termchar *lafter,
   memset(term.post_bidi_cache[line].forward, 0, width * sizeof(int));
   memset(term.post_bidi_cache[line].backward, 0, width * sizeof(int));
 
+  int ib = 0;
   for (i = 0; i < width; i++) {
-    int p = wcTo[i].index;
+    while (wcTo[ib].index == -1)
+      ib++;
+
+    int p = wcTo[ib].index;
 
     assert(0 <= p && p < width);
 
     term.post_bidi_cache[line].backward[i] = p;
     term.post_bidi_cache[line].forward[p] = i;
+
+    ib++;
   }
+  (void)bidisize;
+  assert(ib == bidisize);
 }
 
 #ifdef debug_bidi
@@ -883,7 +892,7 @@ term_bidi_line(termline *line, int scr_y)
     return null;
 
   termchar *lchars;
-  int it;
+  int it, ib;
 
  /* Do Arabic shaping and bidi. */
 
@@ -895,6 +904,7 @@ term_bidi_line(termline *line, int scr_y)
       term.wcTo = renewn(term.wcTo, term.wcFromTo_size);
     }
 
+    ib = 0;
     for (it = 0; it < term.cols; it++) {
       ucschar c = line->chars[it].chr;
 
@@ -908,14 +918,39 @@ term_bidi_line(termline *line, int scr_y)
         }
       }
 
-      term.wcFrom[it].origwc = term.wcFrom[it].wc = c;
-      term.wcFrom[it].index = it;
+      if (it) {
+        termchar * bp = &line->chars[it - 1];
+        // Unfold directional formatting characters which are handled 
+        // like combining characters in the mintty structures 
+        // (and would thus stay hidden from minibidi), and need to be 
+        // exposed as separate characters for the minibidi algorithm
+        while (bp->cc_next) {
+          bp += bp->cc_next;
+          if (bp->chr == 0x200E || bp->chr == 0x200F
+              || (bp->chr >= 0x202A && bp->chr <= 0x202E)
+              || (bp->chr >= 0x2066 && bp->chr <= 0x2069)
+             )
+          {
+            term.wcFromTo_size++;
+            term.wcFrom = renewn(term.wcFrom, term.wcFromTo_size);
+            term.wcTo = renewn(term.wcTo, term.wcFromTo_size);
+            term.wcFrom[ib].origwc = term.wcFrom[ib].wc = bp->chr;
+            term.wcFrom[ib].index = -1;
+            ib++;
+          }
+        }
+      }
+
+      term.wcFrom[ib].origwc = term.wcFrom[ib].wc = c;
+      term.wcFrom[ib].index = it;
+
+      ib++;
     }
 
     trace_bidi("=", term.wcFrom);
-    do_bidi(term.wcFrom, term.cols);
+    do_bidi(term.wcFrom, ib);
     trace_bidi(":", term.wcFrom);
-    do_shape(term.wcFrom, term.wcTo, term.cols);
+    do_shape(term.wcFrom, term.wcTo, ib);
     trace_bidi("~", term.wcTo);
 
     if (term.ltemp_size < line->size) {
@@ -925,16 +960,22 @@ term_bidi_line(termline *line, int scr_y)
 
     memcpy(term.ltemp, line->chars, line->size * sizeof(termchar));
 
+    ib = 0;
     for (it = 0; it < term.cols; it++) {
-      term.ltemp[it] = line->chars[term.wcTo[it].index];
-      if (term.ltemp[it].cc_next)
-        term.ltemp[it].cc_next -= it - term.wcTo[it].index;
+      while (term.wcTo[ib].index == -1)
+        ib++;
 
-      if (term.wcTo[it].origwc != term.wcTo[it].wc)
-        term.ltemp[it].chr = term.wcTo[it].wc;
+      term.ltemp[it] = line->chars[term.wcTo[ib].index];
+      if (term.ltemp[it].cc_next)
+        term.ltemp[it].cc_next -= it - term.wcTo[ib].index;
+
+      if (term.wcTo[ib].origwc != term.wcTo[ib].wc)
+        term.ltemp[it].chr = term.wcTo[ib].wc;
+
+      ib++;
     }
     term_bidi_cache_store(scr_y, line->chars, term.ltemp, term.wcTo,
-                          term.cols, line->size);
+                          term.cols, line->size, ib);
 
     lchars = term.ltemp;
   }